patch RotateFFT; have to restrict to even images for now
