+++ cryojax/ndimage/_fourier_utils.py
@@ -191,3 +191,45 @@ def query_efficient_grid_size(
         return candidate
 
     return tuple(next_smooth_int(n, nf) for n, nf in zip(shape, padded_shape))
+
+
+def make_fftshift_phase(
+    shape: tuple[int, ...],
+    axes: tuple[int, ...] | None = None,
+    outputs_rfft: bool = False,
+) -> Array:
+    """Build the `(-1)^(k1+k2+...)` sign pattern for the full (non-real) FFT.
+
+    Multiplying `jnp.fft.fftn(x)` by this pattern gives the same result as
+    `jnp.fft.fftn(jnp.fft.ifftshift(x))`, so the output has DC at corner
+    (modeord=0 convention). Then `jnp.fft.fftshift` moves DC back to center
+    if needed for storage.
+
+    Only exact for even-sized dimensions.
+
+    **Arguments:**
+
+    - `shape`: shape of the FFT output array (same as input when s=None).
+    - `axes`: axes to include; defaults to all axes.
+
+    **Returns:**
+
+    Broadcastable ±1 array with the same number of dimensions as `shape`.
+    """
+    ndim = len(shape)
+    if axes is None:
+        axes = tuple(range(ndim))
+    else:
+        axes = tuple(ax % ndim for ax in axes)
+    phase = jnp.ones(())
+    for idx, ax in enumerate(axes):
+        n = shape[ax]
+        if outputs_rfft and idx == len(axes) - 1:
+            indices = jnp.arange(n // 2 + 1)
+        else:
+            indices = jnp.arange(n)
+        p = jnp.where(indices % 2 == 0, 1.0, -1.0)
+        reshape = [1] * ndim
+        reshape[ax] = n
+        phase = phase * p.reshape(reshape)
+    return phase
